chore: moved icons to lib/Icons
