+++ src/lib/Icons/Directory.svelte
