+++ src/lib/Icons/File.svelte
